Replaced EntityTagger tests with EntityTaggerDB tests; replaced EntityExplainer tests with EntityExplainerDB tests
