+++ src/nitests/src/frontend/test_entity_explanation.py
@@ -1,15 +1,129 @@
-import string
 from unittest import TestCase
 
+import tqdm
+from sqlalchemy import insert, delete
+
+from narraint.backend.database import SessionExtended
+from narraint.backend.models import EntityExplainerData, IndexVersion, EntityTaggerData
 from narraint.frontend.entity.entityexplainer import EntityExplainer
-from narrant.config import PLANT_GENUS_DATABASE_FILE
-from narrant.vocabularies.excipient_vocabulary import ExcipientVocabulary
+from narraint.frontend.entity.entitytagger import EntityTagger
+
+tagger_entries = [
+    ('CHEMBL4204794', 'Drug', None, ' avapritinib'),
+    ('CHEMBL1064', 'Drug', None, ' simvastatin hydroxy acid'),
+    ('CHEMBL465617', 'Drug', None, ' amantadine sulfate'),
+    ('CHEMBL1569', 'Drug', None, ' amantadine hcl'),
+    ('CVCL:B6WA', 'CellLine', None, ' 1e5 mouse hybridoma against amantadine'),
+    ('CHEMBL1431', 'Drug', None, ' metformin extended release'),
+    ('CHEMBL1431', 'Drug', None, ' la 6023'),
+    ('CHEMBL660', 'Drug', None, ' amantadine'),
+    ('CHEMBL2348413', 'Drug', None, ' metformin pregabalin salt'),
+    ('CHEMBL1569', 'Drug', None, ' amantadine hydrochloride'),
+    ('CHEMBL503', 'Drug', None, ' simvastatin impurity lovastatin'),
+    ('CHEMBL2165491', 'Drug', None, ' spiroamantadine'),
+    ('CHEMBL2108299', 'Drug', None, ' metformin glycinate'),
+    ('CHEMBL494397', 'Drug', None, ' metformin xr'),
+    ('CHEMBL1201391', 'Drug', None, ' simvastatin acid'),
+    ('CHEMBL1703', 'Drug', None, ' la 6023'),
+    ('CHEMBL2348410', 'Drug', None, ' metformin gabapentin salt'),
+    ('CHEMBL1703', 'Drug', None, ' metformin hydrochloride'),
+    ('CHEMBL1703', 'Drug', None, ' metformin hcl'),
+    ('CHEMBL1201391', 'Drug', None, ' simvastatin hydroxy acid'),
+    ('CHEMBL1064', 'Drug', None, ' simvastatin'),
+    ('CHEMBL1431', 'Drug', None, ' metformin'),
+    ('CHEMBL1201391', 'Drug', None, ' simvastatin carboxylic acid'),
+    ('MESH:D000094024', 'Disease', None, ' long haul covid'),
+    ('MESH:C535564', 'Disease', None, ' tibia absence of with polydactyly'),
+    ('MESH:D000094024', 'Disease', None, ' longcovid'),
+    ('CHEMBL486174', 'Drug', None,  ' variotin'),
+    ('MESH:D000094024', 'Disease', None, ' long haul covid19s'),
+    ('MESH:D000094024', 'Disease', None, ' covid19 long haul'),
+    ('MESH:C535563', 'Disease', None, ' absence of tibia'),
+    ('MESH:C564764', 'Disease', None, ' tibia absence of with congenital deafness'),
+    ('MESH:C535563', 'Disease', None, ' tibia absence of'),
+    ('MESH:C535563', 'Disease', None, ' bilateral absence of the tibia'),
+    ('MESH:D000094024', 'Disease', None, ' long covid'),
+    ('CHEMBL486174', 'Drug', None, ' pecilocin'),
+    ('MESH:D000094024', 'Disease', None, ' longhaul covid'),
+    ('MESH:D000094024', 'Disease', None, ' covid longhaul'),
+    ('Q97154236', 'Vaccine', None, ' anhui zhifei longcom biopharmaceutical covid19 vaccine candidate'),
+    ('MESH:C535564', 'Disease', None, ' absence of tibia with polydactyly'),
+    ('MESH:C535689', 'Disease', None, ' fibula and ulna duplication of with absence of tibia and radius'),
+    ('MESH:D000094024', 'Disease', None, ' long haul covid 19'),
+    ('MESH:D000094024', 'Disease', None, ' long hauler covid'),
+    ('MESH:D000094024', 'Disease', None, ' longhaul covids'),
+    ('MESH:D000094024', 'Disease', None, ' long haul covid19'),
+    ('MESH:C563403', 'Disease', None,
+     ' tibia absence or hypoplasia of with polydactyly retrocerebellar arachnoid cyst and other anomalies')
+]
+
+explanation_entries = [
+    ('CHEMBL1201391', '[simvastatin acid,tenivastatin,simvastatin hydroxy acid,simvastatin carboxylic acid]'),
+    ('CHEMBL2165491', '[spiroamantadine]'),
+    ('CHEMBL2348413', '[metformin pregabalin salt]'),
+    ('CVCL:B6WA', '[1e5 [mouse hybridoma against amantadine]]'),
+    ('CHEMBL4204794', '[avapritinib,x-720776,c-366,70c366,blu-285,x720776]'),
+    ('CHEMBL1703',
+     '[ex404,walaphage,benofomin,diabefagos,neodipa,siamformet,diabex,ex-404,apophage,metformin hcl,la-6023,nsc-91485,metformin hydrochloride,glucaminol]'),
+    ('CHEMBL503',
+     '[simvastatin impurity, lovastatin-,sivlor,mevlor,c10aa02,lovastatin,mevinolin,mk-803,nsc-758662,6.alpha.-methylcompactin,l-154803,monacolin k]'),
+    ('CHEMBL494397', '[metformin xr]'), ('CHEMBL2108299', '[metformin glycinate,dmmet-01]'),
+    ('CHEMBL1569',
+     '[adamantanamine hydrochloride,amantadine hcl,mantadix,amantadine hydrochloride,exp-105-1,osmolex,nsc-83653]'),
+    ('CHEMBL660', '[symmetrel,symadine,nsc-341865,amantadine,tcmdc-125869,amantidine]'),
+    ('CHEMBL465617', '[amantadine sulfate]'),
+    ('CHEMBL1064', '[nsc-758706,synvinolin,mk-733,simvastatin,simvastatin hydroxy acid,c10aa01,mk-0733]'),
+    ('CHEMBL1431', '[la-6023,metformin,metformin extended release]'),
+    ('CHEMBL2348410', '[metformin gabapentin salt]'),
+    ('MESH:D000094024',
+     '[long covid,Long-Haul COVIDs,Long Haul COVID 19,Post-Acute COVID-19 Syndromes,Long Haul COVID,Post-COVID Conditions,persistent covid-19,post-covid syndrome,post-acute covid syndrome,post-coronavirus disease-2019 syndrome,Long Haul COVID-19,long-covid,long-haul covid,long-haul coronavirus disease,post-acute covid19 syndrome,chronic coronavirus disease syndrome,Post Acute Sequelae of SARS CoV 2 Infection,Post-COVID Condition,COVID-19 Post-Acute Sequelae,PASC Post Acute Sequelae of COVID 19,long haul covid,Post-Acute Sequelae of COVID-19,Post-Acute Sequelae of SARS-CoV-2 Infection,Post Acute COVID-19 Syndrome,Post Acute COVID 19 Syndrome,chronic covid syndrome,post-acute sequelae of sars-cov-2 infection,long hauler covid,COVID, Long-Haul,post-coronavirus disease syndrome,Post-Acute COVID-19 Syndrome,Post COVID Conditions,post-covid-19 syndrome,COVID-19 Syndrome, Post-Acute,Long COVID,Post Acute Sequelae of COVID 19,PASC Post Acute Sequelae of COVID-19,pasc,Long Haul COVID-19s,Long-Haul COVID,COVID-19, Long Haul,post-acute covid-19 syndrome,post-acute sequelae of severe acute respiratory syndrome coronavirus 2]'),
+    ('MESH:C535563', '[Tibial hemimelia,Absence of Tibia,Bilateral absence of the tibia,Tibia, absence of]'),
+    ('MESH:C564764', '[Tibia, Absence of, with Congenital Deafness]'),
+    ('MESH:C535689',
+     '[Tetramelic mirror-image polydactyly,Mirror hands and feet with nasal defects,Fibula And Ulna, Duplication Of, With Absence Of Tibia And Radius,Laurin-Sandrow Syndrome, Segmental,Laurin Sandrow syndrome,Fibula ulna duplication tibia radius absence,Mirror-Image Polydactyly,Laurin-Sandrow syndrome,Sandrow syndrome]'),
+    ('MESH:C563403',
+     '[Tibia, Absence or Hypoplasia of, with Polydactyly, Retrocerebellar Arachnoid Cyst, and Other Anomalies]'),
+    ('CHEMBL486174', '[variotin,nsc-291839,pecilocin]'),
+    ('Q97154236',
+     '[rbd-dimer,zf-uz-vac-2001,zifivax,anhui zhifei longcom biopharmaceutical covid-19 vaccine candidate,zf2001,zhongyianke biotech–liaoning maokangyuan biotech covid-19 vaccine]'),
+    ('MESH:C535564',
+     '[Tibia, Absence of, with Polydactyly,Absence of tibia with polydactyly,Polydactyly with absent tibia]')
+]
 
 
 class EntityExplanationTestCase(TestCase):
 
-    def setUp(self) -> None:
-        self.entity_explainer = EntityExplainer()
+    @classmethod
+    def setUpClass(cls) -> None:
+        session = SessionExtended.get()
+
+        # update versions
+        session.execute(delete(IndexVersion).where(IndexVersion.name == EntityTagger.NAME))
+        session.execute(insert(IndexVersion).values(name=EntityTagger.NAME, version=EntityTagger.VERSION))
+
+        session.execute(delete(IndexVersion).where(IndexVersion.name == EntityExplainer.NAME))
+        session.execute(insert(IndexVersion).values(name=EntityExplainer.NAME, version=EntityExplainer.VERSION))
+
+        session.commit()
+
+        entity_tagger_data = list()
+        for ent_id, ent_type, ent_class, synonyms in tagger_entries:
+            entity_tagger_data.append(dict(entity_id=ent_id,
+                                           entity_type=ent_type,
+                                           entity_class=ent_class,
+                                           synonym=synonyms,
+                                           synonym_processed=synonyms))
+
+        entity_explainer_data = list()
+        for ent_id, ent_terms in explanation_entries:
+            entry = dict(entity_id=ent_id, entity_terms=ent_terms)
+            entity_explainer_data.append(entry)
+
+        # update index data
+        EntityTaggerData.bulk_insert_values_into_table(session, entity_tagger_data)
+        EntityExplainerData.bulk_insert_values_into_table(session, entity_explainer_data)
+
+        cls.entity_explainer = EntityExplainer()
 
     def test_chembl_entries(self):
         """
@@ -76,125 +190,8 @@ class EntityExplanationTestCase(TestCase):
         self.assertIn('Ace', f_test)
         self.assertIn('Acetabulum', f_test)
 
-    def test_mesh_entries(self):
-        """
-        Tests whether MeSH entries can be tagged correctly
-            :return:
-        """
-        diabetes_2_names = ['Diabetes Mellitus',
-                            'NIDDM',
-                            'Noninsulin Dependent Diabetes Mellitus']
-
-        terms = self.entity_explainer.explain_entity_str('Diabetes Mellitus', truncate_at_k=1000)
-        for dn in diabetes_2_names:
-            self.assertIn(dn, terms)
-
-        # These names should be not included because Diabetes Mellitus is a prefix of them
-        diabetes_already_included_names = ['Diabetes Mellitus, Adult-Onset',
-                                           'Diabetes Mellitus, Ketosis-Resistant',
-                                           'Diabetes Mellitus, Maturity-Onset',
-                                           'Diabetes Mellitus, Non Insulin Dependent',
-                                           'Diabetes Mellitus, Non-Insulin-Dependent',
-                                           'Diabetes Mellitus, Noninsulin Dependent',
-                                           'Diabetes Mellitus, Noninsulin-Dependent',
-                                           'Diabetes Mellitus, Slow-Onset',
-                                           'Diabetes Mellitus, Stable',
-                                           'Diabetes Mellitus, Type II', ]
-        for dn in diabetes_already_included_names:
-            self.assertNotIn(dn, terms)
-
-        neoplasms_terms = ['Neoplasms',
-                           'Benign Neoplasms',
-                           'Cancer',
-                           'Malignancy',
-                           'Malignant Neoplasms',
-                           'Neoplasia',
-                           'Neoplasm',
-                           'Tumors']
-
-        terms = self.entity_explainer.explain_entity_str('Cancer', truncate_at_k=10000)
-        for nt in neoplasms_terms:
-            self.assertIn(nt, terms)
-
-        # Neoplasm is a prefix, thus it should not be included
-        self.assertNotIn('Neoplasms, Benign', terms)
-
     def test_mesh_supplement_entries(self):
         self.assertIn('Absence of Tibia', [t for t in self.entity_explainer.explain_entity_str('Absence of Tibia')])
-
-    def test_plant_families(self):
-        """
-        Tests whether plant family names can be tagged correctly
-        """
-        plant_families_in_db = []
-        with open(PLANT_GENUS_DATABASE_FILE, 'rt') as f:
-            plant_families_in_db = [line.strip() for line in f]
-
-        for pf in plant_families_in_db:
-            self.assertIn(pf.lower(),
-                          [t.lower() for t in self.entity_explainer.explain_entity_str(pf, truncate_at_k=10000)])
-
-    def test_excipient_names(self):
-        """
-        Tests whether excipient names can be tagged correctly
-        """
-        excipient_names = [n for n in ExcipientVocabulary.read_excipients_names(expand_terms=False)]
-        trans_map = {p: '' for p in string.punctuation}
-        translator = str.maketrans(trans_map)
-        for en in excipient_names:
-            if en.strip() and len(en.strip()) > 1:
-                try:
-                    # ignore punctuation and lower/upper case
-                    self.assertIn(en.translate(translator).lower(),
-                                  [t.translate(translator).lower() for t in
-                                   self.entity_explainer.explain_entity_str(en, truncate_at_k=10000)])
-                except AssertionError:
-                    # Try prefixes - Maybe a shorter version is included
-                    # if a prefix is included everything is fine (one synonym was a prefix of the long version)
-                    result = list([t.translate(translator).lower() for t in
-                                   self.entity_explainer.explain_entity_str(en, truncate_at_k=10000)])
-                    found = False
-                    for pref in self.entity_explainer.get_prefixes(en):
-                        if pref in result:
-                            found = True
-                            break
-                    self.assertTrue(found)
-
-    def test_gene_names(self):
-        cyp3a4_names = ['CYP3A4', 'HLP', 'CP33', 'CP34', 'CYP3A', 'NF-25', 'CYP3A3',
-                        'P450C3', 'CYPIIIA3', 'CYPIIIA4', 'P450PCN1']
-        terms = [t for t in self.entity_explainer.explain_entity_str('CYP3A4', truncate_at_k=1000)]
-        for n in cyp3a4_names:
-            self.assertIn(n, terms)
-
-        mtor_names = ['MTOR', 'mechanistic Target of rapamycin', 'SKS', 'FRAP', 'FRAP1', 'FRAP2', 'RAFT1', 'RAPT1']
-        terms = [t for t in self.entity_explainer.explain_entity_str('MTOR', truncate_at_k=1000)]
-        for n in mtor_names:
-            self.assertIn(n, terms)
-
-        cyp3a5_names = ['CYP3A5', 'cytochrome P450 family 3 subfamily A member 5',
-                        'CP35', 'CYPIIIA5', 'P450PCN3', 'PCN3']
-        terms = [t for t in self.entity_explainer.explain_entity_str('cyp3a5', truncate_at_k=1000)]
-        for n in cyp3a5_names:
-            self.assertIn(n, terms)
-
-    def test_nano_particle(self):
-        names = ['Nanoparticles']
-        for n1 in names:
-            terms = [t for t in self.entity_explainer.explain_entity_str(n1, truncate_at_k=1000)]
-            for n2 in names:
-                self.assertIn(n2, terms)
-
-    def test_explain_concept_prefix_filter(self):
-        headings = self.entity_explainer.explain_entity_str("Diabetes", truncate_at_k=10000)
-        self.assertIn("Diabetes Mellitus", headings)
-        self.assertNotIn("Diabetes Mellitus Type 1", headings)
-        self.assertNotIn("Diabetes Mellitus Type 2", headings)
-
-        headings = self.entity_explainer.explain_entity_str("Covid 19", truncate_at_k=10000)
-        self.assertIn("COVID-19", headings)
-        self.assertNotIn("COVID-19 Drug Treatment", headings)
-        self.assertNotIn("COVID-19 Testing", headings)
 
     def test_no_mesh_expansion(self):
         direct_names = ['Long COVID', "Long Haul COVID-19", "Long-Haul COVID"]
@@ -215,3 +212,54 @@ class EntityExplanationTestCase(TestCase):
                 self.assertIn(n2, terms)
             for not_n2 in not_expanded_names:
                 self.assertNotIn(not_n2, terms)
+
+
+def generate_test_data():
+    tagger_data = set()
+    explainer_data = set()
+
+    entity_tagger = EntityTagger()
+
+    terms_to_explain = [
+        "Metformin",
+        "LA-6023",
+        "Simvastatin",
+        "Amantadine",
+        "Avapritinib",
+
+        "Absence of Tibia",
+
+        'Long COVID',
+        "Long Haul COVID-19",
+        "Long-Haul COVID",
+
+        'Pecilocin',
+        "variotin"
+    ]
+    session = SessionExtended.get()
+
+    for term in tqdm.tqdm(terms_to_explain):
+        for entity in entity_tagger.tag_entity(term):
+            tagger_data.add((entity.entity_id, entity.entity_type, entity.entity_class, entity.entity_name))
+
+            query = session.query(EntityExplainerData.entity_terms)
+            query = query.filter(entity.entity_id == EntityExplainerData.entity_id)
+
+            if query.count() == 0:
+                continue
+
+            terms = query.first()[0]
+            explainer_data.add((entity.entity_id, terms))
+
+    print("TAGGER INDEX DATA")
+    print(",\n".join(str(t) for t in list(tagger_data)))
+
+    print()
+    print()
+
+    print("EXPLAINER INDEX DATA")
+    print(",\n".join(str(t) for t in list(explainer_data)))
+
+
+if __name__ == "__main__":
+    generate_test_data()
